+++ src/main/java/com/vibe_guide/traitlike/services/impl/TraitLikeQueryServiceImpl.java
@@ -1,0 +1,32 @@
+package com.vibe_guide.traitlike.services.impl;
+
+import com.vibe_guide.traitlike.services.TraitLikeQueryService;
+import com.vibe_guide.trait.dtos.TraitResponseDTO;
+import com.vibe_guide.traitlike.repositories.TraitLikeRepository;
+import java.util.List;
+import java.util.UUID;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@RequiredArgsConstructor
+@Service
+public class TraitLikeQueryServiceImpl implements TraitLikeQueryService {
+
+  private final TraitLikeRepository traitLikeRepository;
+
+  /**
+   * Retrieves all traits that a given user has liked for a specific place.
+   *
+   * <p>This method uses a lightweight DTO projection query to avoid loading entire entity graphs,
+   * ensuring optimal performance with no N+1 query issues.
+   *
+   * @param placeId the ID of the place whose traits are being fetched
+   * @param userId the ID of the user who liked those traits
+   * @return a list of {@link TraitResponseDTO} representing the liked traits
+   */
+  @Override
+  public List<TraitResponseDTO> getAllByPlaceIdAndUserId(UUID placeId, UUID userId) {
+    return traitLikeRepository.findAllByUserIdAndPlaceId(userId, placeId);
+  }
+}
+
